backend/fastapi-certificados/requirements.txt
fastapi
uvicorn[standard]
requests
Jinja2
qrcode
python-dotenv
